refactor name files
