+++ repositories/experiment.repository.js
@@ -1,5 +1,5 @@
-const { MongoStorage }= require("../db/MongoStorage");
-const validateDate = require("validate-date");
+const { MongoStorage }= require("../db/mongo.storage");
+const validateDate = require('validate-date');
 
 module.exports = new class ExperimentsRepository extends  MongoStorage {
     constructor() {
